refactor(web): extract useSkillDetail() hook from SkillDetailPage (#453)
SkillDetailPage went from 873 lines down to 625 (−28%). The deferred follow-up from #651: pull every query, mutation, derived value, ref-style state, and handler callback out of the page into a single 'useSkillDetail(idOrName)' hook. The page becomes a thin shell that destructures the hook and wires props into already-extracted subcomponents.

What the hook owns:

- Queries: useSkill, useSkillVersions, useAuditSummaryByVersion, useSkillAuditHistory, useSkillPulls (7d), useSkillPackage (files / contents / raw zip).
- Mutations: useDeleteSkill, useDeleteSkillVersion, useUpdateSkillPackage, useSetVersionDeprecation, useRefreshSkillFromSource, useStartAudit.
- Derived state: isOwner, isAdmin, canManageVersions, latestVersion, viewingLatest, pullCount7d, mergedContents, mergedFiles, versionAudit, versionAuditRunning, ownerDisplayName, ownerAvatarUrl, hasChanges.
- Local UI state: 7 modal/expansion toggles (showDeleteConfirm, showPermissionsModal, showAdvancedModal, showSaveConfirm, showAuditStartedModal, showVersions, showVersionDiff) + skipValidation + editedContents + addedPaths + deletedPaths.
- Handlers (all useCallback): handleVersionChange, handleToggleDeprecation, handleContentChange, handleCreateFile, handleCreateFolder, handleDeleteFile, handleSave, handleDeleteConfirm, handleDownloadPackage, handleStartAudit.
- Side-effect: resets the version-browser + diff toggles when skill.guid flips so navigating between two skills doesn't leak modal state.

What stays on the page:

- The page itself owns navigate + addToast since two inline JSX call sites need them (the hero CTA + the copy-GUID button). These could move into the hook but the call-site clarity wins over hook-bloat.
- The layout shell — breadcrumbs, hero, pulls strip, the 2-col grid with tabs/save/preview, the right rail of cards, the modals.
- Aliases at the top of the function body destructure the hook return so the JSX below stays readable.

Cumulative #453 progress on SkillDetailPage: 1133 → 625 lines (−45%) across #651 + this PR. Still above the 300L target, but the data flow is now cleanly split — every state-management concern lives in the hook, every layout concern stays in the page.

Behavior unchanged; 110 web + 793 backend + 17 sdk tests pass.
